requirements.txt
imutils==0.5.2
opencv-python==4.1.1.26
numpy==1.17.0
requests==2.22.0
PyYAML==5.1.2
paho-mqtt==1.4.0